+++ CHANGELOG.md
@@ -275,6 +275,7 @@
     `config :brando, title_prefix: "My App >> "`
   * Brando.Type.Image now has width and height keys. These are set in the `create_image_sizes` function
   * Brando.Type.File now has mimetype and size keys. These are set in the `create_file_struct` function
+  * Default prod setup for HTTPS.
 
 * Bug fixes
   * Fixed gifsicle thumbnailing